requirements.txt
requests
TikTokLive-API
python-dotenv
Flask
Flask-Cors